changed int to float

+++ old.py
@@ -3,7 +3,7 @@
 import sys
 print ("Adrian is a screub")
 
-out_of = int(sys.argv[1])
+out_of = float(sys.argv[1])
 while 1:
     try:
         line = input()
@@ -11,6 +11,6 @@ while 1:
         continue
     total = 0
     for num in line.split(" "):
-        total += int(num)
+        total += float(num)
     print("Total: %d (%f%%)" % (total, total * (100.0 / out_of)))
 print("exit")
